refactor: code cleanup and sync changes
