--- src/extension/options-page/Empty.tsx
@@ -1,13 +1,0 @@
-import React from 'react'
-
-export default (
-    <div className="empty">
-        <style>
-            {`.empty {
-                background: #232D36 url(https://maskbook.com/img/bg--warai-kamen.svg) fixed repeat;
-                width: 100%;
-                height: calc(100vh - 64px);
-            }`}
-        </style>
-    </div>
-)
