+++ src/extension/options-page/PersonaCard.tsx
@@ -6,7 +6,7 @@ import CardContent from '@material-ui/core/CardContent'
 import Button from '@material-ui/core/Button'
 import Typography from '@material-ui/core/Typography'
 import { Person } from '../../database'
-import { Divider, TextField } from '@material-ui/core'
+import { Divider } from '@material-ui/core'
 import Services from '../service'
 import { definedSocialNetworkUIs } from '../../social-network/ui'
 import { Link } from 'react-router-dom'
