+++ src/stories/OptionsPage.tsx
@@ -4,15 +4,11 @@ import Identity from '../components/Dashboard/Identity'
 import { demoPeople } from './demoPeopleOrGroups'
 import { action } from '@storybook/addon-actions'
 import Dashboard from '../components/Dashboard/Dashboard'
-import Privacy from '../extension/options-page/Privacy'
 import { QrCode } from '../components/shared/qrcode'
 import { text } from '@storybook/addon-knobs'
 import { ExportData } from '../components/MobileImportExport/Export'
 
 storiesOf('Options Page', module)
-    // FIXME:
-    // @ts-ignore
-    .add('Privacy', Privacy)
     .add('QrCode', () => <QrCode text={text('QrCode', 'QrCode')} />)
     .add('ExportData', () => <ExportData />)
